% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_detect.R
\name{copy_detect}
\alias{copy_detect}
\title{copy_detect}
\usage{
copy_detect(filename, copy_number, lower_threshold = 0.4, kmer = 5)
}
\arguments{
\item{filename}{A fasta file containing short reads from a single subset generated by "subset_downsize".}

\item{copy_number}{An interger (e.g. 2,3, or 4) giving the expected number of gene copies in the input file.}

\item{lower_threshold}{A decimal (range: 0.01-0.99) giving the OTU identity cutoff for divisive heirarchical clustering analyses. Default 0.4.}

\item{kmer}{An integer giving the k-mer size used to generate the input matrix for k-means clustering. Default 5.}
}
\value{
A fasta alignment of the gene copies found in the input dataset.
}
\description{
This function separates gene copies from a single subset.
}
