% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_separate.R
\name{copy_separate}
\alias{copy_separate}
\title{copy_separate}
\usage{
copy_separate(
  filename,
  copy_number,
  read_length,
  overlap = 225,
  lower_threshold = 0.4,
  kmer = 5
)
}
\arguments{
\item{filename}{A fasta file containing thousands of short reads that have been mapped to a reference gene.}

\item{copy_number}{An interger (e.g. 2,3, or 4) giving the expected number of gene copies in the input file.}

\item{read_length}{An interger (e.g. 250, or 300) giving the read length of your Next-generation Sequencing data. This method is designed for read length 250bp or longer.}

\item{overlap}{An interger giving how many base pairs of overlap between adjacent subsets. Must be smaller than read_length. Default 225.}

\item{lower_threshold}{A decimal (range: 0.01-0.99) giving the OTU identity cutoff for divisive heirarchical clustering analyses.Default 0.4.}

\item{kmer}{An integer giving the k-mer size used to generate the input matrix for k-means clustering. Default 5.}
}
\value{
A fasta alignment of a small number of overlapping DNA sequences covering the entire length of the target gene. Gene copies can be assembled by moving sequences around.
}
\description{
This function separates two or more gene copies from short Next-Generation Sequencing reads.
}
