% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_validate.R
\name{copy_validate}
\alias{copy_validate}
\title{copy_validate}
\usage{
copy_validate(filename, copy_number, read_length)
}
\arguments{
\item{filename}{A DNA alignment in fasta format that contains sequences of two or more gene copies.}

\item{copy_number}{An integer (e.g. 2,3, or 4) giving the number of gene copies in the input file.}

\item{read_length}{An integer (e.g. 250, or 300) giving the read length of your Next-generation Sequencing data.}
}
\value{
A histogram in pdf format showing the comparison between the distance among neighboring variable sites and the read length.
}
\description{
This function checks if chimeric sequences may have been formed during the assembling of gene copies.
}
