% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_downsize.R
\name{subset_downsize}
\alias{subset_downsize}
\title{subset_downsize}
\usage{
subset_downsize(filename, read_length, overlap)
}
\arguments{
\item{filename}{A fasta file containing thousands of short reads that have been mapped to a reference gene.}

\item{read_length}{An interger (e.g. 250, or 300) giving the read length of your Next-generation sequencing data. This method is designed for read length 250bp or longer.}

\item{overlap}{An interger giving how many base pairs of overlap between adjacent subsets. Must be smaller than "read_length".}
}
\value{
A number of overlapping subsets of the input alignment. More overlap means more subsets, but not fewer reads per subset.
}
\description{
This function subdivides the imported read alignment into subsets and then downsizes each subset by deleting those sequences that have too many gaps or missing data.
}
