% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_assemble.R
\name{copy_assemble}
\alias{copy_assemble}
\title{copy_assemble}
\usage{
copy_assemble(filename, copy_number, verbose = 1)
}
\arguments{
\item{filename}{A fasta alignment of a small number of overlapping DNA sequences (results from "copy_separate") covering the entire length of the target gene. Check the alignment carefully before proceeding.}

\item{copy_number}{An integer (e.g. 2,3, or 4) giving the anticipated number of gene copies. Must be the same value as used for "copy_separate".}

\item{verbose}{Turn on (verbose=1; default) or turn off (verbose=0) the output.}
}
\value{
A fasta alignment of the anticipated number of full-length gene copies.
}
\description{
Assembles a small number of overlapping DNA sequences into their respective gene copies.
}
\examples{
\dontrun{
copy_assemble("inst/extdata/combined_con.fasta",2,1)
}

}
