% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_detect.R
\name{copy_detect}
\alias{copy_detect}
\title{copy_detect}
\usage{
copy_detect(filename, copy_number, verbose = 1)
}
\arguments{
\item{filename}{A fasta file contains short reads from a single subset generated by "subset_downsize".}

\item{copy_number}{An integer (e.g. 2,3, or 4) giving the anticipated number of gene copies in the input file.}

\item{verbose}{Turn on (verbose=1; default) or turn off (verbose=0) the output.}
}
\value{
A fasta alignment of the anticipated number of gene copies.
}
\description{
Separates two or more gene copies from a single subset of short reads.
}
\examples{
\dontrun{
copy_detect("inst/extdata/toysubset.fasta",2,1)
}

}
