% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toydata.R
\docType{data}
\name{toydata}
\alias{toydata}
\title{toydata

Provides small example dataset for major functions. They are located at: /inst/extdata

The file "toydata.fasta" is for functions "copy_separate" and "subset_downsize".
The file "combined_con.txt" is for the function "copy_assemble".
The file "toysubset.fasta" is for the function "copy_detect".
The file "Final_two_copiess.fasta" is for the function "copy_validate".}
\format{
fasta
}
\description{
toydata

Provides small example dataset for major functions. They are located at: /inst/extdata

The file "toydata.fasta" is for functions "copy_separate" and "subset_downsize".
The file "combined_con.txt" is for the function "copy_assemble".
The file "toysubset.fasta" is for the function "copy_detect".
The file "Final_two_copiess.fasta" is for the function "copy_validate".
}
\examples{

}
\references{
The data was collected by the author and is unpublished.
}
\keyword{dataset}
