% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toydata.R
\docType{data}
\name{toydata}
\alias{toydata}
\title{toydata}
\format{
fasta
}
\description{
Provides toydata (in the folder: "data") only for package testing during package submission. The real example dataset is located at: /inst/extdata.
}
\details{
The file "toydata.fasta" is for functions "copy_separate" and "subset_downsize".
The file "combined_con.txt" is for the function "copy_assemble".
The file "toysubset.fasta" is for the function "copy_detect".
The file "Final_two_copiess.fasta" is for the function "copy_validate".
}
\examples{

copy_separate
}
\references{
This data set was collected by the author and is unpublished.
}
\keyword{dataset}
